✏ modify : 여행 완료 수정 | 황신운

+++ Travelance/domain/src/main/java/com/moneyminions/domain/model/traveldone/CategoryExpenseDto.kt
@@ -2,5 +2,5 @@ package com.moneyminions.domain.model.traveldone
 
 data class CategoryExpenseDto(
     val category: String,
-    val percent: Int
+    val percent: Double,
 )
